requirements.txt
flask
pandas
openai
python-dotenv
openpyxl  # For reading Excel files if needed
